refactored contract follow/unfollow into model methods
