+++ purchasing_test/unit/data/test_contracts.py
@@ -14,7 +14,7 @@ from purchasing_test.unit.test_base import BaseTestCase, db
 from purchasing_test.unit.util import (
     insert_a_contract, get_a_property
 )
-from purchasing_test.unit.factories import ContractBaseFactory, UserFactory, ContractPropertyFactory
+from purchasing_test.unit.factories import ContractBaseFactory, UserFactory
 
 class ContractRenewalTest(BaseTestCase):
     def setUp(self):
@@ -156,107 +156,3 @@ class ContractsTest(BaseTestCase):
         insert_a_contract()
 
         self.assertEquals(len(get_all_contracts()), 3)
-
-class ContractObjectTest(TestCase):
-    def setUp(self):
-        super(ContractObjectTest, self).setUp()
-        self.today = datetime.datetime.today()
-        self.tomorrow = datetime.datetime.today() + datetime.timedelta(1)
-        self.yesterday = datetime.datetime.today() - datetime.timedelta(1)
-        self.active_contract = ContractBaseFactory.build(
-            expiration_date=datetime.date.today() + datetime.timedelta(1),
-            is_archived=False, parent_id=None
-        )
-
-    def test_create_new_contract(self):
-        '''Test contract object behaves as expected
-        '''
-        # test that no additional properties works
-        contract_data = dict(
-            description='test'
-        )
-
-        contract = ContractBaseFactory.build(**contract_data)
-        self.assertEquals(contract.description, 'test')
-
-        # test that additional properties also works
-        contract_data_props = dict(
-            description='test2'
-        )
-        contract_with_props = ContractBaseFactory.build(**contract_data_props)
-
-        prop1 = ContractPropertyFactory.build(contract=contract_with_props)
-        prop2 = ContractPropertyFactory.build(contract=contract_with_props)
-
-        self.assertEquals(contract_with_props.description, 'test2')
-        self.assertEquals(len(contract_with_props.properties), 2)
-        self.assertTrue(prop1 in contract_with_props.properties)
-        self.assertTrue(prop2 in contract_with_props.properties)
-
-        # this should fail with non-existing stage
-        contract_data_fails = dict(
-            description='test2',
-            current_stage_id=1
-        )
-        try:
-            ContractBaseFactory.build(**contract_data_fails)
-            # this should never work, so break the tests if we get here
-            self.assertTrue(False)
-        except:
-            self.assertTrue(True)
-
-    def test_active_contract_status(self):
-        '''Test active scout contract status
-        '''
-        self.assertEquals(self.active_contract.scout_contract_status, 'active')
-
-    def test_expired_replaced_contract_status(self):
-        '''Test contract that is expired and replaced
-        '''
-        expired_replaced = ContractBaseFactory.build(
-            expiration_date=self.yesterday, is_archived=True,
-            children=[self.active_contract]
-        )
-        self.assertEquals(expired_replaced.scout_contract_status, 'expired_replaced')
-
-    def test_replaced_contract_status(self):
-        '''Test contract that is replaced
-        '''
-        replaced = ContractBaseFactory.build(
-            expiration_date=self.tomorrow, is_archived=True,
-            children=[self.active_contract]
-        )
-        self.assertEquals(replaced.scout_contract_status, 'replaced')
-
-    def test_expired_contract_status(self):
-        '''Test contract that is expired
-        '''
-        expired = ContractBaseFactory.build(
-            expiration_date=self.yesterday, is_archived=True,
-        )
-        self.assertEquals(expired.scout_contract_status, 'expired')
-
-    def test_archived_contract_status(self):
-        '''Test contract that is archived
-        '''
-        archived = ContractBaseFactory.build(
-            expiration_date=self.tomorrow, is_archived=True,
-        )
-        self.assertEquals(archived.scout_contract_status, 'archived')
-
-    def test_no_expiration_replaced(self):
-        '''Test contract that is replaced but has not expiration date
-        '''
-        replaced = ContractBaseFactory.build(
-            is_archived=True,
-            children=[self.active_contract]
-        )
-        self.assertEquals(replaced.scout_contract_status, 'replaced')
-
-    def test_no_expiration_archived(self):
-        '''Test contract that is archived but has no expiration date
-        '''
-        replaced = ContractBaseFactory.build(
-            is_archived=True,
-        )
-        self.assertEquals(replaced.scout_contract_status, 'archived')
